Add assert to hint users that groups cannot contain promises
