+++ tests/dummy/app/templates/public-pages/docs/groups.hbs
@@ -12,7 +12,8 @@
 
 <p>
   You can nest groups inside groups with absolutely no depth limit, everything else like arrow
-  navigation and filtering will just work.
+  navigation and filtering will just work. However, the options of a group cannot be a
+  promise, it has to be a regular collection.
 </p>
 
 <p>
